requirements.txt
streamlit
numpy
pandas
altair
pillow
matplotlib
seaborn